+++ neutcurve/hillcurve.py
@@ -8,6 +8,7 @@ Defines :class:`HillCurve` for fitting neutralization curves.
 
 import collections
 import math
+import warnings
 
 import matplotlib.pyplot as plt
 
@@ -33,12 +34,12 @@ class HillCurve:
 
     When :math:`t = 1` and :math:`b = 0`, this equation is identical to the
     `Hill curve <https://en.wikipedia.org/wiki/Hill_equation_(biochemistry)>`_,
-    except that we are calcuating the fraction **unbound** rather than
+    except that we are calculating the fraction **unbound** rather than
     the fraction bound.
 
-    In some cases, you may want to fit the fraction bound (or neutralized)
-    rather than the fraction infectivity (or unbound). In that case, set
-    `infectivity_or_bound='bound'` and then the equation that is fit will
+    You may want to fit the fraction neutralized rather than the fraction
+    infectivity. In that case, set `infectivity_or_neutralized='neutralized'
+    and then the equation that is fit will
     be :math:`f\left(c\right) = t + \frac{b - t}{1 + \left(c/m\right)^s}`,
     which means that `f\left(c\right)` gets larger rather than smaller
     as :math:`c` increases.
@@ -48,9 +49,9 @@ class HillCurve:
             Concentrations of antibody / serum.
         `fs` (array-like)
             Fraction infectivity remaining at each concentration.
-        `infectivity_or_bound` ({'infectivity', 'bound'})
-            Fit the fraction infectivity (:math:`f\left(c\right)` decreases
-            as :math:`c` increases), or fraction bound (:math:`f\left(c\right)`
+        `infectivity_or_neutralized` ({'infectivity', 'neutralized'})
+            Fit the fraction infectivity (:math:`f\left(c\right)` decreases as
+            :math:`c` increases) or neutralized (:math:`f\left(c\right)`
             increases as :math:`c` increases). See equations above.
         `fs_stderr` (`None` or array-like)
             If not `None`, standard errors on `fs`.
@@ -257,13 +258,33 @@ class HillCurve:
     >>> numpy.allclose(0.2, neut.fracinfectivity(neut.icXX(0.8)))
     True
 
+    Now fit with `infectivity_or_neutralized='neutralized'`, which is useful
+    when the signal **increases** rather than decreases with increasing
+    concentration (as would be the case if measuring fraction bound rather
+    than fraction infectivity).
+
+    .. nbplot::
+
+       >>> neut_opp = HillCurve(cs, [1 - f for f in fs],
+       ...                      fixtop=False, fixbottom=False,
+       ...                      infectivity_or_neutralized='neutralized')
+       >>> numpy.allclose(neut_opp.top, 0.9)
+       True
+       >>> numpy.allclose(neut_opp.bottom, 0)
+       True
+       >>> numpy.allclose(neut_opp.midpoint, m)
+       True
+       >>> neut_opp.ic50() < neut_opp.midpoint
+       True
+       >>> fig, ax = neut_opp.plot(ylabel='fraction neutralized')
+
     """
 
     def __init__(self,
                  cs,
                  fs,
                  *,
-                 infectivity_or_bound='infectivity',
+                 infectivity_or_neutralized='infectivity',
                  fs_stderr=None,
                  fixbottom=0,
                  fixtop=1,
@@ -282,9 +303,19 @@ class HillCurve:
         self.fs = self.fs[self.cs.argsort()]
         self.cs = self.cs[self.cs.argsort()]
 
-        if infectivity_or_bound not in {'infectivity', 'bound'}:
-            raise ValueError('invalid `infectivity_or_bound`')
-        self._infectivity_or_bound = infectivity_or_bound
+        if infectivity_or_neutralized == 'infectivity':
+            if self.fs[0] < self.fs[-1] and (self.fs[0] < 0.3 and
+                                             self.fs[-1] > 0.7):
+                warnings.warn('`f` increases with concentration, consider '
+                              '`infectivity_or_neutralized="neutralized"')
+        elif infectivity_or_neutralized == 'neutralized':
+            if self.fs[0] > self.fs[-1] and (self.fs[0] > 0.7 and
+                                             self.fs[-1] < 0.3):
+                warnings.warn('`f` decreases with concentration, consider '
+                              '`infectivity_or_neutralized="infectivity"')
+        else:
+            raise ValueError('invalid `infectivity_or_neutralized`')
+        self._infectivity_or_neutralized = infectivity_or_neutralized
 
         if any(self.cs <= 0):
             raise ValueError('concentrations in `cs` must all be > 0')
@@ -360,25 +391,28 @@ class HillCurve:
             initguess = [midpoint, slope, bottom, top]
 
             def func(c, m, s, b, t):
-                return evalfunc(c, m, s, b, t, self._infectivity_or_bound)
+                return evalfunc(c, m, s, b, t,
+                                self._infectivity_or_neutralized)
 
         elif fixtop is False:
             initguess = [midpoint, slope, top]
 
             def func(c, m, s, t):
-                return evalfunc(c, m, s, bottom, t, self._infectivity_or_bound)
+                return evalfunc(c, m, s, bottom, t,
+                                self._infectivity_or_neutralized)
 
         elif fixbottom is False:
             initguess = [midpoint, slope, bottom]
 
             def func(c, m, s, b):
-                return evalfunc(c, m, s, b, top, self._infectivity_or_bound)
+                return evalfunc(c, m, s, b, top,
+                                self._infectivity_or_neutralized)
         else:
             initguess = [midpoint, slope]
 
             def func(c, m, s):
                 return evalfunc(c, m, s, bottom, top,
-                                self._infectivity_or_bound)
+                                self._infectivity_or_neutralized)
 
         (popt, pcov) = scipy.optimize.curve_fit(
                 f=func,
@@ -456,27 +490,30 @@ class HillCurve:
             bounds = bounds + [(None, None), (None, None)]
 
             def func(c, m, s, b, t):
-                return evalfunc(c, m, s, b, t, self._infectivity_or_bound)
+                return evalfunc(c, m, s, b, t,
+                                self._infectivity_or_neutralized)
 
         elif fixtop is False:
             initguess = [midpoint, slope, top]
             bounds.append((bottom, None))
 
             def func(c, m, s, t):
-                return evalfunc(c, m, s, bottom, t, self._infectivity_or_bound)
+                return evalfunc(c, m, s, bottom, t,
+                                self._infectivity_or_neutralized)
 
         elif fixbottom is False:
             initguess = [midpoint, slope, bottom]
             bounds.append((None, top))
 
             def func(c, m, s, b):
-                return evalfunc(c, m, s, b, top, self._infectivity_or_bound)
+                return evalfunc(c, m, s, b, top,
+                                self._infectivity_or_neutralized)
         else:
             initguess = [midpoint, slope]
 
             def func(c, m, s):
                 return evalfunc(c, m, s, bottom, top,
-                                self._infectivity_or_bound)
+                                self._infectivity_or_neutralized)
 
         def min_func(p):
             """Evaluate to zero when perfect fit."""
@@ -628,34 +665,35 @@ class HillCurve:
         """Fraction infectivity at `c` for fitted parameters."""
         return self.evaluate(c, self.midpoint, self.slope,
                              self.bottom, self.top,
-                             self._infectivity_or_bound)
+                             self._infectivity_or_neutralized)
 
     @staticmethod
     def evaluate(c, m, s, b, t,
-                 infectivity_or_bound='infectivity'):
+                 infectivity_or_neutralized='infectivity'):
         r""":math:`f\left(c\right) = b + \frac{t-b}{1+\left(c/m\right)^s}`.
 
-        If `infectivity_or_bound` is 'bound' rather than 'infectivity', instead
+        If `infectivity_or_neutralized` is 'neutralized' rather than
+        'infectivity', instead return
         :math:`f\left(c\right) = t + \frac{b-t}{1+\left(c/m\right)^s}`.
 
         """
-        if infectivity_or_bound == 'infectivity':
+        if infectivity_or_neutralized == 'infectivity':
             return b + (t - b) / (1 + (c / m)**s)
-        elif infectivity_or_bound == 'bound':
+        elif infectivity_or_neutralized == 'neutralized':
             return t + (b - t) / (1 + (c / m)**s)
         else:
-            raise ValueError('invalid `infectivity_or_bound`')
+            raise ValueError('invalid `infectivity_or_neutralized`')
 
     @staticmethod
     def _evaluate_log(logc, logm, s, b, t,
-                      infectivity_or_bound='infectivity'):
+                      infectivity_or_neutralized='infectivity'):
         """Like :class:`HillCurve.evaluate` but on log concentration scale."""
-        if infectivity_or_bound == 'infectivity':
+        if infectivity_or_neutralized == 'infectivity':
             return b + (t - b) / (1 + numpy.exp(s * (logc - logm)))
-        elif infectivity_or_bound == 'bound':
+        elif infectivity_or_neutralized == 'neutralized':
             return t + (b - t) / (1 + numpy.exp(s * (logc - logm)))
         else:
-            raise ValueError('invalid `infectivity_or_bound`')
+            raise ValueError('invalid `infectivity_or_neutralized`')
 
     def plot(self,
              *,